requirements.txt
asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.3
django>=5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-cors-headers==4.3.1
whitenoise==6.6.0
djangorestframework==3.14.0
django-filter==23.5